fix: check delete permissions when cancelling enrollment

Closes LINSELF-47

See merge request dev/linotp/ng-selfservice!107

+++ src/app/enroll/enroll-motp-dialog/enroll-motp-dialog.component.ts
@@ -12,6 +12,7 @@ import { SetPinDialogComponent } from '../../common/set-pin-dialog/set-pin-dialo
 import { TokenTypeDetails, EnrollToken } from '../../api/token';
 import { EnrollmentService } from '../../api/enrollment.service';
 import { OperationsService } from '../../api/operations.service';
+import { NgxPermissionsService } from 'ngx-permissions';
 
 @Component({
   selector: 'app-enroll-motp',
@@ -39,6 +40,7 @@ export class EnrollMOTPDialogComponent implements OnInit {
     public dialogRef: MatDialogRef<EnrollMOTPDialogComponent>,
     @Inject(MAT_DIALOG_DATA) public data: { tokenTypeDetails: TokenTypeDetails, closeLabel: String },
     public notificationService: NotificationService,
+    public permissionsService: NgxPermissionsService,
     public i18n: I18n,
   ) { }
 
@@ -103,7 +105,7 @@ export class EnrollMOTPDialogComponent implements OnInit {
    * Cancel the dialog and return false as result
    */
   public cancelDialog() {
-    if (this.enrolledToken) {
+    if (this.enrolledToken && this.permissionsService.hasPermission(Permission.DELETE)) {
       this.operationsService.deleteToken(this.enrolledToken.serial).subscribe();
     }
     this.dialogRef.close(false);
